requirements.txt
certifi==2018.11.29
dj-database-url==0.5.0
Django==2.1.5
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
Pillow==5.4.1
pipenv==2018.11.26
psycopg2==2.7.7
pytz==2018.9
virtualenv==16.2.0
virtualenv-clone==0.5.1
whitenoise==4.1.2
